renaming and other small changes
